# AK+base64: Performance improvements for base64 encoding and decoding