Using http://mattmahoney.net/dc/enwik8.zip (100MB unzipped) as a test:

Encoding runtime decreases from 1.192s to 0.303s (74.6% faster)
Decoding runtime decreases from 0.917s to 0.469s (48.9% faster)